refactor: mission
