+++ src/app/home.tsx
@@ -1,16 +1,16 @@
 import NoticeBoardAlert from "@/components/NoticeBoardAlert";
 import Hero from "@/components/hero";
-import FeaturedMissions from "@/components/mission/FeaturedMissions";
-import MissionsByCategory from "@/components/mission/MissionsByCategory";
 import RankingSection from "@/components/top/ranking-section";
 import { MetricsWithSuspense } from "@/features/metrics/components/metrics-with-suspense";
+import FeaturedMissions from "@/features/missions/components/featured-missions";
+import MissionsByCategory from "@/features/missions/components/missions-by-category";
+import { hasFeaturedMissions } from "@/features/missions/services/missions";
 import Activities from "@/features/user-activity/components/activities";
 import { BadgeNotificationCheck } from "@/features/user-badges-notification/components/badge-notification-check";
 import { getUnnotifiedBadges } from "@/features/user-badges/services/get-unnotified-badges";
 import { LevelUpCheck } from "@/features/user-level/components/level-up-check";
 import { checkLevelUpNotification } from "@/features/user-level/services/level-up-notification";
 import { generateRootMetadata } from "@/lib/metadata";
-import { hasFeaturedMissions } from "@/lib/services/missions";
 import { getCurrentSeasonId } from "@/lib/services/seasons";
 import { createClient } from "@/lib/supabase/client";
 import { redirect } from "next/navigation";
