+++ src/features/missions/components/mission-card.tsx
@@ -1,14 +1,14 @@
 "use client";
 
+import { Button } from "@/components/ui/button";
 import { Card, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
+import { MissionIcon } from "@/components/ui/mission-icon";
 import { calculateMissionXp } from "@/features/user-level/utils/level-calculator";
 import type { Tables } from "@/lib/types/supabase";
 import clsx from "clsx";
 import { motion } from "framer-motion";
 import { UsersRound } from "lucide-react";
 import Link from "next/link";
-import { Button } from "../ui/button";
-import { MissionIcon } from "../ui/mission-icon";
 import MissionAchievementStatus from "./mission-achievement-status";
 
 interface MissionProps {
