+++ src/features/missions/components/mission-details.tsx
