+++ src/features/missions/components/mission-guidance-arrow.tsx
