+++ src/features/missions/components/poster-form.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import {
@@ -9,11 +10,10 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
+import { Separator } from "@/components/ui/separator";
 import { Textarea } from "@/components/ui/textarea";
 import { VALID_JP_PREFECTURES } from "@/features/map-poster/constants/poster-prefectures";
 import { useState } from "react";
-import { Button } from "../ui/button";
-import { Separator } from "../ui/separator";
 
 type PosterFormProps = {
   disabled: boolean;
